requirements.txt
pillow
opencv-python
openpyxl
natsort